Conflicts:
	ui/scripts/ui/dialog.js

+++ server/src/com/cloud/template/HypervisorTemplateAdapter.java
@@ -215,11 +215,15 @@ public class HypervisorTemplateAdapter extends TemplateAdapterBase implements Te
                     } else {
                         _tmpltHostDao.remove(templateHostVO.getId());
                     }
-                    VMTemplateZoneVO templateZone = _tmpltZoneDao.findByZoneTemplate(sZoneId, templateId);
-
-                    if (templateZone != null) {
-                        _tmpltZoneDao.remove(templateZone.getId());
+                    //check if the list is not null and empty,check to see if there is a entry and remove it.
+                    List<VMTemplateZoneVO> templateZones = _tmpltZoneDao.listByZoneTemplate(sZoneId, templateId);
+                    if (templateZones != null && templateZones.size()!=0) {
+                        VMTemplateZoneVO  templateZone = templateZones.get(0);
+                        if (templateZone != null) {
+                            _tmpltZoneDao.remove(templateZone.getId());
+                        }
                     }
+
                 } finally {
                     if (lock != null) {
                         _tmpltHostDao.releaseFromLockTable(lock.getId());
